[ATR] Introduce ProtoComposer concept
PrefixSplice is now replaced by ScopedPrefix which needs to be called in
the Recompose function instead of directly calling PrefixSplice. This simplifies
the concept and wrapping of a Composer. However, for now the syntax appears
more complex in the Recompose function. The added albeit yet unused advantage
is, that after a ScopedPrefix multiple Splices, Members, or Aliases with
the same prefix may occur.

Against expectations the build duration did not suffer in a meaningful way
because of the concept.