[Evaluation] Update compile times
The procedure is now as follows: After every relevant change, the minimum
observed time will be updated. The script should be run until the first
run where no time is lower than a previous run. This should yield the times
the most independent of outside factors. It is also easier to compare iterations
than some value or even an average of multiple runs, which may have different
CPU loads at the time. It represents the "ideal" time.