[Evaluation] Improve handling of non power of 2 reductions
Previously when using TransformReduce with 2^16 + 7 elements, the sum for
the first 2^16 would be computed and added to the sum of the last 7. For
floating point numbers this is undesirable and defeats the purpose of this
algorithm over std::transform_reduce. Now the sum of elements 0-6 will be
added to the sum of 7-14 and 15-22, from which point on the summation continues
as normal.

Added a test case based on integers to verify that the algorithm is correct.
Removed artificial SIMD for non-SIMD iterators. The comparison should be
multiple elements vs single element not some single/multiple hybrid