[Evaluation] Use constructors for initialization
This change was long overdue. It greatly improves readability. Due to the
additional code required, this changes compile times negatively for most
cases, but appears to have a positive effect on runtime.

ATR however benefits from this change in terms of compile time, as there
is no longer the need for the (yet unoptimized) member access and aggregate
initialization can be used for the most part.

In a future version, ATR constructors should become order-independent with
names attached to each argument, but for simplicity sake they are not right
now.

This places TagATR at the top for compile speed of all cases with any layer
of abstraction.