[ATR] Fix district index for more than 1 district
Due to left to right evaluation of operator or, the right to left fold expression
did not work as intended:

(a or ...)
evaluates to
(a2 or (a1 or a0))
which evaluates a1 before a0 and, in particular even if a0 is true. Therefore
a fold expression (a or ...  or false) would evaluate a0 before a1 but would
still evaluate a1.